requirements.txt
discord.py
pip
random